requirements.txt
# Core ML and Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.0.76
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2

# Deep Learning
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
easyocr==1.7.0

# MLOps Tools
mlflow==2.7.1
dvc==3.20.1
hydra-core==1.3.2
wandb==0.15.8

# API and Web Framework
fastapi==0.103.1
uvicorn==0.23.2
gradio==3.40.1
streamlit==1.26.0

# Google Cloud
google-cloud-vision==3.4.4
google-cloud-storage==2.10.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
requests==2.31.0

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1
pre-commit==3.4.0

# Monitoring and Logging
prometheus-client==0.17.1
structlog==23.1.0

# Data Processing
albumentations==1.3.1
imgaug==0.4.0 